+++ src/web/helpers.js
@@ -219,7 +219,8 @@ local.web.preferredType = function preferredType(accept, provided) {
 // </https://github.com/federomero/negotiator>
 
 // EXPORTED
-// correctly joins together to url segments
+// correctly joins together all url segments given in the arguments
+// eg joinUrl('/foo/', '/bar', '/baz/') -> '/foo/bar/baz/'
 local.web.joinUrl = function joinUrl() {
 	var parts = Array.prototype.map.call(arguments, function(arg, i) {
 		arg = ''+arg;
@@ -230,6 +231,31 @@ local.web.joinUrl = function joinUrl() {
 		return arg.substring(lo, hi);
 	});
 	return parts.join('/');
+};
+
+// EXPORTED
+// takes a context url and a relative path and forms a new valid url
+// eg joinRelPath('http://grimwire.com/foo/bar', '../fuz/bar') -> 'http://grimwire.com/foo/fuz/bar'
+local.web.joinRelPath = function(urld, relpath) {
+	if (typeof urld == 'string')
+		urld = local.web.parseUri(urld);
+	if (relpath.charAt(0) == '/')
+		// "absolute" relative, easy stuff
+		return urld.protocol + '://' + urld.authority + relpath;
+	// totally relative, oh god
+	// (thanks to geoff parker for this)
+	var hostpath = urld.path;
+	var hostpathParts = hostpath.split('/');
+	var relpathParts = relpath.split('/');
+	for (var i=0, ii=relpathParts.length; i < ii; i++) {
+		if (relpathParts[i] == '.')
+			continue; // noop
+		if (relpathParts[i] == '..')
+			hostpathParts.pop();
+		else
+			hostpathParts.push(relpathParts[i]);
+	}
+	return local.web.joinUrl(urld.protocol + '://' + urld.authority, hostpathParts.join('/'));
 };
 
 // EXPORTED